refactor: 멤버별 예약 조회 책임 Reservation 도메인으로 이동
